[Ref]  Extracting `PerfCompareTests`
Refactoring - Extracting `PerfCompareTests` for the performance comparison test, priorly included in the CLI project
